Major Refactoring of codebase introduction of docstrings.
