+++ dasc2/bin/build_orders.py
@@ -1,123 +1,161 @@
 #!/usr/bin/env python
 #
-# Portions of this code inspired by:
-# https://github.com/cole-maclean/autocraft/tree/master/autocraft/EDA/Build%20Orders
+# Copyright 2017 David Kuhta & Anshul Sacheti. All Rights Reserved.
 #
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
 #
-# The above copyright notice and this permission notice shall be included in
-# all copies or substantial portions of the Software.
+#      http://www.apache.org/licenses/LICENSE-2.0
 #
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-# THE SOFTWARE.
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS-IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 
-import json
 import os
-import time
-import ast
+import json
+import argparse
+import collections
 
-from mpyq import MPQArchive
 from pkg_resources import resource_filename
 
-def build_order(replay_data, build_orders_dir):
-   unit_ids = build_unit_dict()
+def create_unit_dict():
+    """Constructs a dictionary of unit_ids to unit_labels
+        
+        Returns:
+            Dictionary: { unit_id : unit_name } <int> : <string>
+    """
 
-   '''Script to parse out the build orders from a single state_data file.
-   Build orders are the unique units seen from player1's persepective for both players
-   in the order they are seen in the replay.
-   '''
-   build_order = []
-   friendly_build = []
-   enemy_build = []
-   #iterate over state data and find unique units from friendly and enemy player through the states
-   for state in replay_data:
-       state_build = []
-       #load all friendly units seen in this state
-       for unit_data in state['friendly_army']:
-           #check if unit count > 0
-           if unit_data[1] > 0:
-               unit = unit_ids[unit_data[0]]
-               #lookup unit_id name and append to build if we haven't
-               #seen this unit in previous states
-               if unit not in friendly_build:
-                   friendly_build.append(unit)
-                   #append player marker (friendly = 0) to identify player's unit
-                   #in build order
-                   state_build = state_build + [unit + str(0)]
-       for unit_data in state['enemy_army']:
-           if unit_data[1] > 0:
-               unit = unit_ids[unit_data[0]]
-               if unit not in enemy_build:
-                   enemy_build.append(unit)
-                   state_build = state_build + [unit + str(1)]
-       #update build order if new unit seen this state for either player
-       if state_build:
-           build_order = build_order + state_build
-   #gather static data from  first state
-   player_id = replay_data[0]['player_id']
-   winner = replay_data[0]['winner']
-   if player_id == winner:
-       won = True
-   else:
-       won = False
-   race = replay_data[0]['race']
-   enemy_race = replay_data[0]['enemy_race']
-   game_map = replay_data[0]['map_name']
-   replay_id = replay_data[0]['replay_id']
-   build_data = { "build_order" : build_order, "won" : won, "race" : race,
-                  "enemy_race" : enemy_race, "game_map" : game_map,
-                  "replay_id" : replay_id }
-
-   json_filename = os.path.join(build_orders_dir, "Build_Orders_" + replay_id[0:20] + "_" + str(player_id) + '.json')
-   with open(json_filename, 'w') as json_file:
-       json.dump(build_data, json_file)
-
-def format_states_data(states_data):
-    text_file = open(states_data,"r")
-    lines=text_file.readlines()
-    replay_processed = []
-    for line in lines:
-        replay = ast.literal_eval(line)
-        replay_processed.append(replay)
-    print(len(replay_processed))
-    return replay_processed
-
-def build_unit_dict():
     with open(resource_filename(__name__, '../ref/units.json')) as units_data:
         units = json.load(units_data)
         unit_ids = { int(unit_id) : name for unit_id, name in units.items()}
     return unit_ids
-#     units_json = json.loads(units_file)
-#     units_dict = {int()}
 
-def build_orders(states_dir, build_orders_dir='./build_orders'):
-    # Create directory for replays if it does not exist
-    if not os.path.exists(build_orders_dir):
-        os.makedirs(build_orders_dir)
+def units_seen(states, army_type):
+    """Identifies units seen, adding to dictionary in the order seen
 
+        Args:
+            states: list of game states.
+            army_type: string "Commanding" or "Opposing".
+        
+        Returns:
+            Dictionary: { unit_id : step_when_seen } <int> : <int>
+    """
+    units_seen_dict = collections.OrderedDict()
+    for state in states:
+        for unit in state['Armies'][army_type]:
+            if not unit in units_seen_dict.keys():
+                units_seen_dict[unit] = state['Step']
+
+    return units_seen_dict
+
+def label_units(units_dict, unit_id_dict):
+    """Remaps the keys of dictionary to their label in another dictionary
+
+        Args:
+            units_dict: Dictionary of unit_ids to steps_when_seen.
+            unit_id_dict: Dictionary of unit_ids to unit_labels.
+        
+        Returns:
+            Dictionary: { unit_id : step_when_seen } <int> : <int>
+
+        Example:
+        With a dictionary of {... 43 : "Viking" ... }
+        we would remap a units_dict of:
+            { ... 43 : 800 ... } to { ... "Viking" : 800 ... }
+    """
+    labeled_units_seen_dict = collections.OrderedDict()
+    for unit_id, step in units_dict.items():
+        label = unit_id_dict[int(unit_id)]
+        labeled_units_seen_dict[label] = step
+
+    return labeled_units_seen_dict
+
+def build_order(states, unit_id_dict):
+    """Generates the build data for one replay
+
+        Args:
+            states: list of states for a particular replay.
+            unit_id_dict: dictionary to use for relabeling.
+        
+        Returns:
+            Dictionary: { "Commanding" : ..., "Opposing" : ... } <string> : <list>
+    """
+    commanding_b_o = units_seen(states, 'Commanding')
+    opposing_b_o = units_seen(states, 'Opposing')
+    labeled_commanding_b_o = label_units(commanding_b_o, unit_id_dict)
+    labeled_opposing_b_o = label_units(opposing_b_o, unit_id_dict)
+
+    build_data = { "Commanding " : labeled_commanding_b_o.keys(),
+                    "Opposing" : labeled_opposing_b_o.keys() }
+
+    return build_data
+
+def build_orders(states_dir='./states', build_orders_dir='./build_orders'):
+    """Process as state files in a directory, generating build orders for each
+
+        Args:
+            states_dir: Directory containing a list of states.
+            build_orders_dir: Directory to output build orders.
+    """
+
+    # If states directory exists
     if os.path.exists(states_dir):
-        state_files = []
 
+        # Create directory for replays if it does not exist
+        if not os.path.exists(build_orders_dir):
+            os.makedirs(build_orders_dir)
+
+        state_files = []
+        unit_id_dict = create_unit_dict()
+
+        # Iterate through files in states directory evaluating
+        # only those that are json and appending to a running list
         for file in os.listdir(states_dir):
             if file.endswith(".json"):
                 state_files.append(os.path.join(states_dir, file))
 
+            # We'll try to use tqdm for pretty printing a status bar
+            # which will succeed if the user selected to install it
+            # during the `pip install dasc2`
+            try:
+                from tqdm import tqdm
+                state_files = tqdm(state_files)
+            except ImportError:
+                pass
+
+        # Iterate through the list of state files
         for states_file in state_files:
-            formatted_states_file = format_states_data(states_file)
-            build_order(formatted_states_file, build_orders_dir)
+
+            # load the replay state data as a JSON object
+            replay_state_data = json.load(open(states_file))
+
+            # pass the ['States'] element to the build_order function
+            build_data = build_order(replay_state_data['States'], unit_id_dict)
+
+            # Duplicate the JSON state data object
+            bo_data = replay_state_data
+
+            # Remove the ['States'] element as we've already processed the data
+            del bo_data['States']
+
+            # Inject a new ['ArmyBuildOrder'] element using the processed build data
+            bo_data['ArmyBuildOrder'] = build_data
+
+            # Construct a new JSON file
+            states_filename = os.path.basename(states_file)
+            bo_filename = os.path.join(build_orders_dir, "BO_" + states_filename)
+
+            # Write out the JSON file
+            with open(bo_filename, 'w') as json_file:
+                json.dump(bo_data, json_file)
+                json_file.write('\n')
 
 def parse_args():
-    import argparse
+    """Helper function to parse dasc2_build_order arguments
+    """
 
     parser = argparse.ArgumentParser()
     parser.add_argument('--states_dir', dest='s_dir', action='store',
