+++ dasc2/bin/download_replays.py
@@ -1,7 +1,5 @@
 #!/usr/bin/env python
 #
-# Portions of code reused from following link in accordance with license below:
-# https://github.com/Blizzard/s2client-proto/blob/master/samples/replay-api/download_replays.py
 # Copyright (c) 2017 Blizzard Entertainment
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
@@ -21,6 +19,9 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
+#
+# Portions of code reused from following link in accordance with license above:
+# https://github.com/Blizzard/s2client-proto/blob/master/samples/replay-api/download_replays.py
 
 import os
 import requests
@@ -29,24 +30,11 @@ import argparse
 import urlparse
 import shutil
 
-from pkg_resources import resource_filename
+from dasc2.lib.replay_helpers import check_build_version
 
 API_BASE_URL = 'https://us.api.battle.net'
 API_NAMESPACE = 's2-client-replays'
 
-def check_build_version(build_version):
-    version_file = open(resource_filename(__name__, '../ref/build_versions.json')).read()
-    versions = json.loads(version_file)
-
-    # Check if supplied build matches a build version label
-    if any(v['label'] == str(build_version) for v in versions):
-        return build_version
-
-    # Check if supplied build version matches a 'version'
-    build_label = next((v['label'] for v in versions if (v['version'] == int(build_version))), None)
-    if build_label:
-        return build_label
-    raise ValueError('Build version not supported')
 
 def get_bnet_oauth_access_token(url, key, secret):
     headers = { "Content-Type": "application/json"}
@@ -60,6 +48,7 @@ def get_bnet_oauth_access_token(url, key, secret):
     if 'access_token' in response:
         return response['access_token']
     raise Exception('Failed to get oauth access token. response={}'.format(response))
+
 
 def get_base_url(access_token):
     headers = {"Authorization": "Bearer " + access_token}
@@ -138,6 +127,10 @@ def get_replay_pack(client_version, client_key, client_secret, output_dir, extra
         files = []
 
         sorted_urls = sorted(download_urls)
+
+        # We'll try to use tqdm for pretty printing a status bar
+        # which will succeed if the user selected to install it
+        # during the `pip install dasc2`
         try:
             from tqdm import tqdm
             sorted_urls = tqdm(sorted_urls)
@@ -154,6 +147,8 @@ def get_replay_pack(client_version, client_key, client_secret, output_dir, extra
 
 
 def parse_args():
+    """Helper function to parse dasc2_download arguments
+    """
     parser = argparse.ArgumentParser()
     parser.add_argument('--key', dest='client_key', action='store', type=str,
                         help='Battle.net API key', required=True)
@@ -165,10 +160,11 @@ def parse_args():
                         help='the directory where the downloaded replay archives will be saved to')
     return parser.parse_args()
 
+
 def main():
     args = parse_args()
 
-    processed_version = check_build_version(args.s2_client_version)
+    processed_version = str(check_build_version(args.s2_client_version, True))
 
     get_replay_pack(processed_version, args.client_key, args.client_secret, args.a_dir)
