+++ dasc2/bin/filter_replays.py
@@ -1,24 +1,18 @@
 #!/usr/bin/env python
 #
-# Copyright (c) 2017 David Kuhta & Anshul Sacheti
+# Copyright 2017 David Kuhta & Anshul Sacheti. All Rights Reserved.
 #
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
 #
-# The above copyright notice and this permission notice shall be included in
-# all copies or substantial portions of the Software.
+#      http://www.apache.org/licenses/LICENSE-2.0
 #
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-# THE SOFTWARE.
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS-IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 
 import json
 import os
@@ -27,80 +21,88 @@ import time
 from mpyq import MPQArchive
 
 from dasc2.lib.replay_filter import SC2ReplayFilter
+from dasc2.lib.replay_helpers import check_build_version
 
-def filter_replays(replay_filter, replays_dir='./replays',
-                    filters_dir='./filters', full_path=False):
+def filter_replays(replay_filter):
 
-    replays_dir = os.path.abspath(replays_dir)
-    filters_dir = os.path.abspath(filters_dir)
-
-    if not os.path.exists(filters_dir):
-        os.makedirs(filters_dir)
-
-    metadata_dir = filters_dir + "/metadata"
-    if not os.path.exists(metadata_dir):
-        os.makedirs(metadata_dir)
-    replay_filter._info()
-
-    time_string = "_" + time.strftime("%Y%m%d-%H%M%S")
-    mmr_string = "_MMR_" + str(replay_filter.mmr_threshold)
-    apm_string = "_APM_" + str(replay_filter.apm_threshold)
-
-    filtered_name = "Filtered_Replays"+ mmr_string + apm_string + time_string
-    json_filename = os.path.join(metadata_dir, filtered_name + ".json")
-    txt_filename =  os.path.join(filters_dir, filtered_name + ".txt")
+    replays_dir = os.path.abspath(replay_filter.replays_dir)
 
     total_replay_counter = 0
     filtered_replay_counter = 0
 
-    with open(json_filename, 'w') as json_file, open(txt_filename, 'w') as txt_file:
-        replay_metadata_list = []
+    replay_list = []
+    replay_metadata_list = []
 
-        file_list = os.listdir(replays_dir)
-        PROGRESS_BAR_FLAG = False
+    file_list = os.listdir(replays_dir)
 
-        try:
-            from tqdm import tqdm
-            file_list = tqdm(file_list)
-            PROGRESS_BAR_FLAG = True
-        except ImportError:
-            pass
+    try:
+        from tqdm import tqdm
+        file_list = tqdm(file_list)
+    except ImportError:
+        pass
 
-        for filename in file_list:
+    for file in file_list:
 
-            if filename.endswith(".SC2Replay"):
-                total_replay_counter += 1
-                replay = os.path.join(replays_dir, filename)
-                replay_metadata = replay_filter.filter(replay)
+        if file.endswith(".SC2Replay"):
+            total_replay_counter += 1
+            replay = os.path.join(replays_dir, file)
+            replay_metadata = replay_filter.filter(replay)
 
-                if replay_metadata:
-                    filtered_replay_counter += 1
-                    if full_path:
-                        txt_file.write(replay+"\n")
-                    else:
-                        txt_file.write(filename+"\n")
-                    replay_json = {str(filename.split(".")[0]):replay_metadata}
-                    replay_metadata_list.append(replay_json)
-
-        json.dump(replay_metadata_list, json_file)
+            if replay_metadata:
+                filtered_replay_counter += 1
+                filename = os.path.splitext(file)[0]
+                replay_metadata_list.append({ filename : replay_metadata })
 
     if total_replay_counter:
         print(str(filtered_replay_counter) + " of " + str(total_replay_counter) + " replays met filtering criteria\n")
     else:
         print("No SC2 Replays to evaluate in this directory")
 
-    if filtered_replay_counter:
-        print("List of replays written to: " + str(txt_filename))
-        print("JSON data for replays written to: " + str(json_filename))
-    else:
-        os.remove(json_filename)
-        os.remove(txt_filename)
+    return replay_metadata_list
+
+def generate_filter_file(replay_filter, replay_list):
+    
+    filter_dir = os.path.abspath(replay_filter.filter_dir)
+
+    if not os.path.exists(filter_dir):
+        os.makedirs(filter_dir)
+
+    json_filename = os.path.join(filter_dir, replay_filter.name + "_filter" + ".json")
+
+    with open(json_filename, 'w') as json_file:
+
+        json_output = {}
+
+        json_output.update({ 'Name' : replay_filter.name})
+        json_output.update({ 'CreatedAt' : time.strftime("%Y%m%d-%H%M%S")})
+        json_output.update({ 'DataBuild' : replay_filter.game_version})
+
+        if replay_filter.map_title:
+            map_title = replay_filter.map_title
+        else:
+            map_title = 'all'
+
+        json_output.update({ 'Maps' : map_title})
+
+
+        json_output.update({ 'MinMMR' : replay_filter.mmr_threshold })
+        json_output.update({ 'MinAPM' : replay_filter.apm_threshold })
+        json_output.update({ 'Races' : replay_filter.races })
+        json_output.update({ 'WinningRaces' : replay_filter.winning_races })
+        json_output.update({ 'Replays' : replay_list })
+        json_output.update({ 'ReplaysDirectory' : os.path.abspath(replay_filter.replays_dir) })
+
+        json.dump(json_output, json_file)
+
+        print("Replay Filter generated")
 
 
 def parse_args():
     import argparse
 
     parser = argparse.ArgumentParser()
+    parser.add_argument('--name', dest='name', action='store', default='SC2Filter',
+                        help='Filter Name', required=False)
     parser.add_argument('--min_mmr', dest='mmr', action='store', default=1000,
                         type=int, help='Minimum Player MMR', required=False)
     parser.add_argument('--min_apm', dest='apm', action='store', default=10,
@@ -128,10 +130,17 @@ def parse_args():
 
 def main():
     args = parse_args()
-    replay_filter = SC2ReplayFilter(args.mmr, args.apm, args.races, args.winning_races,
-                        args.map, args.vers)
 
-    filter_replays(replay_filter, args.r_dir, args.f_dir, args.full_path)
+    processed_version = int(check_build_version(args.vers, False))
+
+    replay_filter = SC2ReplayFilter(args.mmr, args.apm, args.races, args.winning_races,
+                        args.map, processed_version, args.r_dir, args.f_dir, args.name)
+
+    replay_filter._info()
+
+    filtered_replay_list = filter_replays(replay_filter)
+    if filtered_replay_list:
+        generate_filter_file(replay_filter, filtered_replay_list)
 
 if __name__ == '__main__':
     main()
