+++ dasc2/bin/move_replays.py
@@ -1,14 +1,25 @@
 #!/usr/bin/env python
+#
+# Copyright 2017 David Kuhta & Anshul Sacheti. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS-IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 
 import os
+import argparse
+
 from shutil import copyfile
 
-from dasc2.lib.replay_filter import SC2ReplayFilter
-
-
 def parse_args():
-    import argparse
-
     parser = argparse.ArgumentParser()
     parser.add_argument('--replayFileList', dest='rFile', action='store', default=None, help='File containing list of replays',required=True)
     parser.add_argument('--outputDir', dest='outputDir', action='store', default=None, help='Directory to copy replays to',required=True)
