+++ dasc2/lib/__init__.py
@@ -1,4 +1,4 @@
-# Copyright 2017 David Kuhta. All Rights Reserved.
+# Copyright 2017 David Kuhta & Anshul Sacheti. All Rights Reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
